requirements.txt
appdirs==1.4.3
argcomplete==1.10.3
certifi==2019.11.28
chardet==3.0.4
idna==2.8
Pillow==6.2.1
pycparser==2.19
PyMuPDF==1.16.9
PyQt5==5.13.2
PyQt5-sip==12.7.0
PyQt5-stubs==5.13.1.3
pyserial==3.4
python-escpos==2.2.0
pyusb==1.0.2
qrcode==6.1
requests==2.22.0
six==1.13.0
urllib3==1.25.7
